requirements.txt
boto3
google-cloud-storage
google-cloud-iam
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib